requirements.txt
pandas==1.1.5; python_version < '3.7'
pandas==1.3.5; python_version == '3.7'
pandas==2.0.1; python_version > '3.7'
google-cloud-aiplatform[pipelines]>=1.25.0
fastapi
uvicorn[standard]